1- and 2-simplex cases, computing normal direction to origin
These functions return the vector pointing towards the origin on a 1- or
2-simplex, while being normal to the nearest vertex/line (or face for
3/4 simplices) to the origin on the simplex. The 1- and 2- simplex case
are probably much simpler than the 3- and 4-simplex cases, since they do
not involve any conditional logic. (Well technically they do involve
conditional logic, but it can be ignored since these are internal to the
gjk algorithm, and the other cases can be ruled out).