Implemented the function for finding direction normal to 4-simplex pointing to origin.
This is a complicated function with many cases, and it is possible that
some of them are not actually possible to reach, based on how this
function is called in GJK.

I have not tested this function yet. To save time, I am moving first to
end-to-end tests of GJK intersection tests, rather than testing the
internal components.