Added collision detection to the demo application.
Collisions are indicated by highlighting cubes red.

I identified and fixed 1 bug in my gjk algorithm. The number of points
in the simplex was changing but not being updated. There are probably
several other bugs - a test is failing and the algorithm will sometimes
get caught in an infinite loop. Will fix these later.