Added GJK iteration cap and stats output
The demo application is showing a bug in GJK, where the algorithm
doesn't terminate. This commit adds an optional parameter to
gjk_intersect limiting the number of iterations, and an optional
statistics output returning the number of itertations. The demo
application then checks this to determine if the algorithm terminated.