More conventional implementation of simplex3_dir
Since I am seeing bugs in the GJK implementation, I want to be confident
in the implementation of simplex3_dir before debugging simplex4_dir.

Before this change, the implementation assumed that it was impossible
for the origin to be in region E (as in the comment at the beginning of
the function). This is something I have derived on paper, but I am not
100% confident in my derivation, and as far as I know this is not
commonly assumed. E.g. Casey Muratori (who summarized this algorithm
very well) did not know of any implementations using this few branches
as of 2016 (see https://caseymuratori.com/blog_0003). So I modified the
logic to assume it is possible for the origin to be in region E, which I
think is a safer assumption even though I have proved otherwise.