Fixed multiple bugs in simplex4_dir, now working reliably.
The major bug is that I was not updating the simplex vertices before returning,
but this is very important since every return condition results in a different
subset of the vertices composing the simplex.

The second bug was one of the plane tests testing the wrong plane.

I knew there were bugs since the demo application was failing: sometimes GJK
would not terminate in a reasonable number of iterations, other times there
would be intersections when there weren't supposed to be. I identified what
the bugs were by careful inspection of the code. It now seems to be working
reliably.