Refactored keyboard input code
Added a helper class which allows keyboard events to be registered and
called. This separates some of the logic for keyboard input out of the
main loop, however it is not super useful for all of the keyboard input:
it is easier to keep all the object position inputs as they are, because
moving them into lambdas requires the check for object existence to be
duplicated in every lambda, which is error prone. I am hoping to come up
with a better solution.