GJK interface change: Support functions are template arguments
Instead of passing the support functions as std::functions, they are
passed as template parameters. This might be slightly faster? The
reasoning was knowing the address at compile time would allow the branch
address to be predicted statically. But now that I am thinking about it,
the support functions are called in a loop so the branch location will
stay in the icache. So there would probably be no practical difference
with a lambda/capture approach. I'll change it to that in the next
commit.