More simplifications to the main loop
Wrapping of the selected_object and selected_mesh indices is handled in
the keyboard event handlers, so this doesn't need to happen separately
in the main loop.

Also, the mesh of the selected object now always updates to the selected
mesh. This is a slightly different design than before - originally, the
selected mesh only applied to the next object created. But this was not
the most natural design choice: switching meshes with the arrow keys was
still relative to the selected mesh, and there was no way to update the
selected object to the selected mesh. Now the selected object always has
the selected mesh. If the user wants to make an object with a different
mesh, they must first create a new object, then change the selected
mesh.