Updated collision detection behaviour to match README
The readme says that every pair of objects is tested for intersection,
not just testing every object with the selected object. I originally
didn't want to do this to avoid the n^2 complexity of testing every pair
of objects, but really the number of objects will be far too small for
that to matter here.