Fixed two bugs with selected object index.
First bug: Deleting the last object in the list of objects would result
in an invalid selected object ID. Fix: After an object is deleted,
select the first object.

Second bug: Modulo was not working correctly for negative dividends.
Since one argument of "%" was signed and the other was unsigned, the
signed was probably getting converted to unsigned which is wrong. So now
the unsigned is being cast to signed.